requirements.txt
Flask==2.3.3
pandas==2.1.1
openpyxl==3.1.2
python-docx==0.8.11
mammoth==1.6.0
python-dotenv==1.0.1
Werkzeug==2.3.7
sendgrid==6.11.0
twilio==8.2.0
fpdf2
